docker: Disable quay.io caching
quay.io never updates our containers unless we change the list of
packages, so we don't get updates when new packages are released, and we
fail to detect regressions in packages we depend on.

Add a version to the RUN line. The version should be modified each time
we want to rebuild the containers.

Adding a workflow the build the container is probably better, but
require more work.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>