backends: Extract Closed helper to common module
We want to have the same close behavior in all backends. Extract the
Closed helper class to a new common module to be used by all backends.

The Closed class fixed to raise ValueError instead of IOError. This is
the behavior of io.FileIO that we use in other modules.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>