logging: Replace ticket id with transfer id
Use the public transfer id instead of the sensitive ticket id used for
authentication. This makes it easier to follow transfer by grepping the
transfer id.

The ticket id is till logged when adding a ticket, we may want to remove
it from the log, or log only a part of the ticket.

Fixes #26

Signed-off-by: Nir Soffer <nsoffer@redhat.com>