bench: Start benchmarks directory
The directory includes imageio configuration and ticket for starting
imageio server for benchmarks purpose, and upload benchmark using
external storage.

We need to add a script to create standard images for benchmarking, so
developers or testers only need to provide storage for the benchmarks.

The benchmarks should work with local directory, NFS, and iSCSI/FC LUN.

We need a light-weight version of the benchmarks that can run in the CI
environment.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>