docker: Add python 3.8 to el8 container
We want to build also python 3.8 version of the common and client
packages, needed by ovirt python sdk. Adding python 3.8 to the container
enables the build and the tests.

python38-systemd is not available, but this is not an issue since we can
install it using pip. To make it work, we need the systemd-devel
package.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>