tox: Fix tests with multiple python versions
Change tox to work with multiple python versions, so we can run the    
tests with both platform python (e.g. 3.6 on centos 8) and python 3.8,  
required for ovirt python sdk.

Looks like python setup.py build_ext --inplace is broken now, and the
built extension is copied to the current directory instead of into    
ovirt_imageio/_internal package. Fixed by using --build-lib option.

The extension is build now in commands_pre, running since it is not
really part of the test but preparation for the test.

The use_develop option is enabled to avoid the annoying step of
installing the ovirt_imageio package in .tox/{env}/lib, which works only
if the ioutil extension module was built before running tox. With
use_develop, tox create a link:

$ cat .tox/test-py38/lib64/python3.8/site-packages/ovirt-imageio.egg-link
/home/nsoffer/src/ovirt-imageio

So we use the source directory for the tests, which contains the ioutil
extension.

In setup.py we use now setuptools instead of distutils, required for
use_develop. This change is also wanted regardless of python 3.8
support, since distutils is deprecated in python 3.10 and we we get
warnings during build when using it.

Finally, install the systemd module using tox for the python 3.8 build,
since the python3-systemd package is built only for platform python. We
install systemd only for the python 3.8 build, since it fail to install
with pip when using python 3.10.

With these changes, we can install python 3.8 on Fedora 35, or in the
test container, and the tests run with both platform python and python
3.8:

$ tox
...
  flake8: commands succeeded
SKIPPED:  test-py36: InterpreterNotFound: python3.6
SKIPPED:  test-py37: InterpreterNotFound: python3.7
  test-py38: commands succeeded
SKIPPED:  test-py39: InterpreterNotFound: python3.9
  test-py310: commands succeeded
SKIPPED:  bench-py36: InterpreterNotFound: python3.6
SKIPPED:  bench-py37: InterpreterNotFound: python3.7
  bench-py38: commands succeeded
SKIPPED:  bench-py39: InterpreterNotFound: python3.9
  bench-py310: commands succeeded
  congratulations :)

Signed-off-by: Nir Soffer <nsoffer@redhat.com>