ci: Don't run make
There is nothing to make before the tests. Running make is recommended
for developers to be able to run the daemon directly from source, but
the tests create the C extension when needed automatically.

Added a new section to docs/developement.md showing how to build and run
the daemon directly from source.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>