tests: Create base image in Makefile
The base image used for backup tests was created by a pytest fixture.
This is useful since the image is created transparently when needed, but
in CI environment creating the image can be very slow and requires very
long timeouts to avoid random failures.

Move the creation of the image to new "images" target. The "check"
target depends on it so the image will be create automatically when
running "make check".

With this change the backup tests run successfully on all tested
distros.  Previously they were skipped on centos 9 since they always
timed out.

Testing shows that the backup tests take 19-23 seconds on github, so 60
seconds timeout should be enough.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>