spec: Add python3-setuptools build requirement
The Fedora 34 and 35 builds in copr fails now with:

+ /usr/bin/python3 setup.py build '--executable=/usr/bin/python3 -s'
Traceback (most recent call last):
  File "/builddir/build/BUILD/ovirt-imageio-2.4.3/setup.py", line 9, in <module>
    from setuptools import setup
ModuleNotFoundError: No module named 'setuptools'

And indeed python3-setuptools is not installed during this build.

The CentOS builds work, and we can see in the logs that
platform-python-setuptools is installed.

Add explicit requirement for python3-setuptools to make sure we have the
package even if someone changes the dependencies.

Fixes: 4ab3229b55bc (tox: Fix tests with multiple python versions)
Signed-off-by: Nir Soffer <nsoffer@redhat.com>