container: Add ovirt-imageio container
Start a container image for running imageio and container.

Using a container, the user will run an image transfer pod for every
transfer, exposing the pod for remote access, and connecting the
relevant storage the pod. When the transfer is done, the user will
unexpose the pod and terminate it.

In this mode the imageio server will serve one ticket, injected into the
pod when starting the pod, instead of adding a ticket using the control
service.

To created the container run:

    make container

This creates localhost/ovirt-imageio container image using podman
configured for rootless environment[1].

To run the container:

    podman run --publish 8000:8000 ovirt-imageio

More work is needed:
- Inject a ticket and PKI files into the container
- Mount the image to transfer or nbd socket in the container
- Modify the server to load a ticket from file during startup

[1] https://github.com/containers/podman/blob/main/docs/tutorials/rootless_tutorial.md

Signed-off-by: Nir Soffer <nsoffer@redhat.com>