container: Add ovirt-imageio container
Start a container image for running imageio and container.

Using a container, the user will run an image transfer pod for every
transfer, exposing the pod for remote access, and connecting the
relevant storage the pod. When the transfer is done, the user will
unexpose the pod and terminate it.

In this mode the imageio server will serve one ticket, injected into the
pod when starting the pod, instead of adding a ticket using the control
service.

To created the container run:

    make container

This creates localhost/ovirt-imageio container image using podman
configured for rootless environment[1].

To run the container:

    podman run --publish 8080:80 -v /path/to/image:/images:Z ovirt-imageio

Next we only need to load a ticket and create the image in the mounted
images folder. Note that the path for the file in the ticket needs
to match the path where we mounted the images folder in the container.

Alternatively, we can mount the nbd socket path in the container
and request the image with a nbd ticket.

More work is needed:
- Modify the server to load a ticket from file during startup
- Document how to mount the image to transfer or the nbd socket in
  the container.
- Inject PKI files into the container (can be done later)

[1] https://github.com/containers/podman/blob/main/docs/tutorials/rootless_tutorial.md

Signed-off-by: Nir Soffer <nsoffer@redhat.com>