Rewrite release script
Previously if "git describe" failed, the error was ignored silently and
we treated the build as build from tag. This creates artifacts with the
wrong release string, breaking users using the copr repo like OST.

Fixed by rewriting in python which is easier to get right and nicer to
work with.

With this change if a git command fails, the build will fail loudly
instead of creating wrong results.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>