Always build ioutil for platform python
When running py38 tests, we run ovirt-imageio using /usr/bin/python3,
but we built ioutil C extension with "python", which is python 3.8. We
cannot fix this by running the ovirt-imageio with python 3.8 since we
don't support python 3.8 for for the server.

Change tox to build the C extension for both current python version and
platform python.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>