admin_tool_test: add wait at srv fixture
In admin_tool_test, srv fixture is parametrized
to create two servers with built ovirt-imageio,
with a terminate at the end, between runs.

However, if runs do happen too fast
it may try to create the next server socket
before the previous is closed, resulting in
an "address already in use" error.

Add a wait() after terminate() to ensure we do
not run into such issue.

Signed-off-by: Albert Esteve <aesteve@redhat.com>