tests: avoid empty socket
A recent change in cpython broke autobind of abstract
unix socket in linux, causing bind calls with an
empty string to be bound to '\0' instead of a
random address (expected). This results in an
'address already in use' error when consecutive empty
string binding is used in the tests.

While this issue may get solved eventually, we
can change the test strategy and use the random
unique addresses in the abstract space, to create
unique sockets per test so that they do not collide,
and do not depend on the empty socket feature.

Related: https://github.com/python/cpython/issues/94821
Signed-off-by: Albert Esteve <aesteve@redhat.com>