uhttp: force empty string to random addr
A recent change in cpython broke autobind of abstract
unix socket in linux, causing bind calls with an
empty string to be bound to '\0' instead of a
random address (expected). This results in an
'address already in use' error when consecutive empty
string binding is used.

While we wait for this issue to be solved, we need
to enforce a random address to be used for empty
strings in the abstract space.

Related: python/cpython#94821
Signed-off-by: Albert Esteve <aesteve@redhat.com>