ci: separate envs per distro
On centos 8 we run the tests twice both with python 3.6
(platform python) and python 3.8. This doubles the
run time and makes it harder to check failures.
Other distros only run a single python environment
tests, but checks for all of them, creating a lot
of unnecesary noise in the logs.

Separate centos-8 enviornments for py36 and py38
to allow parallel execution. Other distributions
should run only a specific python environment
that matches the one they have installed.

Fixes: #90
Signed-off-by: Albert Esteve <aesteve@redhat.com>