ovirt-img: Add configuration file
Add -c,--config option specifying the config section in the ovirt-img
configuration file.

Example file:

    $ cat ~/.config/ovirt-img.conf
    [engine]
    engine_url = https://my.engine
    username = admin@internal
    password = password
    cafile = /home/user/certs/my.engine.pem

The configuration file can contain multiple sections. This can be useful
for people managing multiple oVirt environments.

Using a configuration file downloading is much simpler:

    $ ./ovirt-img download-disk -c engine adb9b2c7-32a9-4e87-894c-710de7165086 disk.qcow2

If the same options are specified both in the command line and the
config file, command line options win.

The password option is special case; if --password-file is specified, we
read the password from the password file even if password is specified
in the config file.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>