ovirt-img: Fix running without arguments
When running without arguments we fail after parsing the arguments
because `config` is not set, and required arguments are missing. Fix by
detecting the case when no command was specified and failing with useful
help (like ovirt-img --help). This is a common behavior for tool that
has sub commands (e.g git).

Signed-off-by: Nir Soffer <nsoffer@redhat.com>