client: Show progress phase
Transferring an image include several phases that may take several
seconds, and we want to show what the tool is doing. When we finish, we
want to show that the operation was completed. If we failed, we want to
show a short failure message even if we use log errors to file.

Add an optional phase argument to ProgressBar. If the phase is set, it
is displayed at the end:

    [ ---- ] 0 bytes, 0.00 s, 0 bytes/s | setting up

The caller can change the phase by setting a new value:

    pb.phase = "downloading image"

This will redraw the progress bar:

    [   0% ] 0 bytes, 0.00 s, 0 bytes/s | downloading image

This change only adds the infrastructure, no change yet in the actual
commands.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>