client: Optimize upload to new image
When we upload to a new disk using qcow2 format or new file we know that
the disk is empty and the entire disk contents is read as zeros by the
guest. In this case we can skip zero extents on the source image instead
of zeroing them on the destination image.

Zeroing on the destination image is usually fast, but when uploading big
images it can slow down the upload a lot. When uploading to qcow2 image,
zeroing creates suboptimal image that will be slower to read and copy
later.

Example upload of empty 8 TiB image:

Before:

    ./ovirt-img upload-disk -c engine --storage-domain fc-01 empty-8t.qcow2
    [ 100% ] 8.00 TiB, 222.82 s, 36.77 GiB/s | upload completed

After:

    ./ovirt-img upload-disk -c engine --storage-domain fc-01 empty-8t.qcow2
    [ 100% ] 8.00 TiB, 11.51 s, 711.88 GiB/s | upload completed

Example upload of 8 TiB Fedora 35 image:

Before:

    $ ./ovirt-img upload-disk -c engine --storage-domain fc-01 fedora-35-8t.qcow2
    [ 100% ] 8.00 TiB, 317.88 s, 25.77 GiB/s | upload completed

After:

    $ ./ovirt-img upload-disk -c engine --storage-domain fc-01 fedora-35-8t.qcow2
    [ 100% ] 8.00 TiB, 109.13 s, 75.07 GiB/s | upload completed

Fixes: #76
Signed-off-by: Nir Soffer <nsoffer@redhat.com>