http_test: make realistic server url
Recent HTTPRequestHandler changes avoid urls that
start with '//' to protect against open redirect
attacks (see https://github.com/python/cpython/commit/4abab6b603dd38bec1168e9a37c40a48ec89508e).

Server url in the http test have no path,
therefore the handler that is used in the
test receives the root path, resulting in
requests starting with '//' when the Backend
tries to get the extents, which break in the
pipelines for recent Python 3.10 versions.

This is not a realistic scenario and not
something the library needs to handle.
Is better to change the url to a more
realistic scenario, and since we want
to test image handling, have
`/images/ticket-id` in the url path.

Signed-off-by: Albert Esteve <aesteve@redhat.com>