ovirt-img: create disk check ILLEGAL
On disk-upload, ovirt-img  probes a disk after
creation during a fixed time, until it reaches an
OK status (or timeout).

If there is an error in the creation and disk becomes
DiskStatus.ILLEGAL, users have to wait for the timeout.

We need to handle this error status and fail loudly.

Fixes: #117
Signed-off-by: Albert Esteve <aesteve@redhat.com>