tests: avoid Backend usage when not needed
Currently all tests that rely on userstorage employ
the Backend wrapper only to add a couple attributes.

With recent update, context management is natively
supported by userstorage library. On the other
hand, not all test require the Backend additional
attributes.

We can avoid using the Backend wrapper completely
for tests that do not require the attributes.
Otherwise, the Backend can just wrap the
storage from within the userstorage context (i.e.,
after the storage is setup), so we avoid the
__enter__() and __exit__() overrides.

Signed-off-by: Albert Esteve <aesteve@redhat.com>