container: add ticket and image
Add image and the associated ticket to the
container so that the containerized imageio
can serve it.

Image is generated and copied to the container
at build time during `make container`.
It creates a fedora-36 image in qcow2 format,
so that it serves a realistic image with
meaningful contents.

Copying the image into the container increases
its size, but allows it to have a default
ticket with well-known size, format, and name.

Signed-off-by: Albert Esteve <aesteve@redhat.com>