reuse: init reuse folder
Do reuse init folder with GPLv2+ license
and custom template for copyright.
This is the main, and default license
for this project.

With reuse all copyright headers will be changed
to SPDX format, while taking these considerations:
- Avoid years in the copyright lines to ease maintenance
- Avoid duplication, remove current copyright headers
- Keep copyright holder intact

Note that copyright notices are optional for works
published after March 1, 1989. On top of that,
dating in copyright notices is usually a guideline
in most licenses. For GPLv2, dating appears after
END OF TERMS AND CONDITIONS, under the
"How to Apply These Terms to Your New Program",
clearly left as a mere guideline.
So, while we should still push to include copyright
notices in every file, it is fine to skip the year.

Signed-off-by: Albert Esteve <aesteve@redhat.com>