reuse: addheader *.pem
Add SPDX header to certificate example files in
the project. We could add the header to the files,
as long as it remains out of the BEGIN ---- END
block. However, since the certificates get
generated through test/genpki script, headers
would get lost if we re-generate them.

Instead of modifying the script to artificially
add the headers (which defies the purpose of the
script in the first place), add explicit license
files, that will work even if we replace the files.

Signed-off-by: Albert Esteve <aesteve@redhat.com>