auth: support tickets with name
Current "uuid" key in the ticket dictionary
is misleading, as it does not necessarily
need to be an uuid, it is just a string to
identify the ticket.

We need to keep it for back-compatibility,
as other projects use the "uuid" key and
there is much code involved.

We can, instead, support "name" key in the
ticket dictionary, as an alternative to
"uuid". So we leave the "uuid" as a legacy
key, but still supported. In the ticket,
if not "uuid" nor "name" are set, we will
raise asking for "name", as is the new
key. If both are set, "uuid" prevails.

Signed-off-by: Albert Esteve <aesteve@redhat.com>