auth_test: test ticket identifiers
Add optional "name" argument to the
create_ticket helper function in the
testutil module, similar to "uuid".

If no "uuid" or "name" are set, the
helper function sets "uuid" to keep
current test expectations stable, but
if we give a "name", it prevails and
no "uuid" is set in the ticket.

This way, we can add a few tests in
auth_test module to check that both
"uuid" and "name" keys work as expected.

Signed-off-by: Albert Esteve <aesteve@redhat.com>