tox.ini: remove ioutil platform build
Remove ioutil platform build from tox
pre-environment setup commands. It is
only needed in the virtual environment.

Otherwise it fails during setup with:
`failed with /usr/bin/python3 (resolves to /usr/bin/python3)
is not allowed, use allowlist_externals to allow it`

The platform build was introduced in the commit b76751a.
Since platform python (/usr/bin/python3) and
"python" were referring to different Python
versions, we needed to build ioutil for both
to run tests.

Currently this is not required anymore, as
we run tests inside the virtual environment,
and all links point to the same version.

Signed-off-by: Albert Esteve <aesteve@redhat.com>